+++ hivemq-edge-pulse-integration-api/src/main/java/com/hivemq/edge/pulse/integration/api/asset/Asset.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.pulse.asset;
+package com.hivemq.edge.pulse.integration.api.asset;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
