+++ hivemq-edge-pulse-integration-api/src/main/java/com/hivemq/edge/pulse/integration/api/management/PulseAgentStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.pulse.status;
+package com.hivemq.edge.pulse.integration.api.management;
 
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
