+++ hivemq-edge-pulse-integration-api/src/main/java/com/hivemq/edge/pulse/integration/api/message/IncomingMessage.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.edge.pulse.integration.api;
+package com.hivemq.edge.pulse.integration.api.message;
 
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
