+++ hivemq-edge/src/main/java/com/hivemq/api/resources/impl/PulseApiImpl.java
@@ -64,9 +64,9 @@ import com.hivemq.pulse.converters.PulseAgentAssetMappingStatusConverter;
 import com.hivemq.pulse.converters.PulseAgentAssetSchemaConverter;
 import com.hivemq.pulse.converters.PulseAgentAssetsConverter;
 import com.hivemq.pulse.converters.PulseAgentStatusConverter;
-import com.hivemq.edge.pulse.integration.api.PulseManagement;
+import com.hivemq.edge.pulse.integration.api.management.PulseManagement;
 import com.hivemq.pulse.management.PulseManagementHolder;
-import com.hivemq.pulse.status.PulseAgentStatus;
+import com.hivemq.edge.pulse.integration.api.management.PulseAgentStatus;
 import com.hivemq.pulse.utils.PulseAgentAssetUtils;
 import com.hivemq.util.ErrorResponseUtil;
 import jakarta.inject.Inject;
