+++ hivemq-edge/src/main/java/com/hivemq/pulse/asset/AssetFactoryImpl.java
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 package com.hivemq.pulse.asset;
+import com.hivemq.edge.pulse.integration.api.asset.Asset;
+import com.hivemq.edge.pulse.integration.api.asset.AssetFactory;
 
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
