+++ hivemq-edge/src/main/java/com/hivemq/pulse/asset/AssetProviderRegistryImpl.java
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 package com.hivemq.pulse.asset;
+import com.hivemq.edge.pulse.integration.api.asset.AssetProviderRegistry;
+import com.hivemq.edge.pulse.integration.api.asset.ExternalAssetProvider;
 
 import com.hivemq.configuration.reader.AssetMappingExtractor;
 import com.hivemq.configuration.reader.PulseExtractor;
