+++ hivemq-edge/src/main/java/com/hivemq/pulse/utils/PulseAgentAssetUtils.java
@@ -23,7 +23,7 @@ import com.hivemq.configuration.entity.pulse.PulseAssetsEntity;
 import com.hivemq.configuration.entity.pulse.PulseEntity;
 import com.hivemq.configuration.reader.AssetMappingExtractor;
 import com.hivemq.configuration.reader.PulseExtractor;
-import com.hivemq.pulse.asset.Asset;
+import com.hivemq.edge.pulse.integration.api.asset.Asset;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
