+++ hivemq-edge/src/test/java/com/hivemq/api/resources/impl/pulse/AbstractPulseApiImplTest.java
@@ -35,9 +35,9 @@ import com.hivemq.edge.api.model.EntityReference;
 import com.hivemq.edge.api.model.EntityReferenceList;
 import com.hivemq.edge.api.model.EntityType;
 import com.hivemq.edge.api.model.Instruction;
-import com.hivemq.edge.pulse.integration.api.PulseManagement;
+import com.hivemq.edge.pulse.integration.api.management.PulseManagement;
 import com.hivemq.pulse.management.PulseManagementHolder;
-import com.hivemq.pulse.status.PulseAgentStatus;
+import com.hivemq.edge.pulse.integration.api.management.PulseAgentStatus;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
