+++ hivemq-edge/src/test/java/com/hivemq/api/resources/impl/pulse/PulseApiImplGetManagedAssetsTest.java
@@ -28,7 +28,7 @@ import com.hivemq.pulse.asset.PulseAgentAssetMapping;
 import com.hivemq.pulse.asset.PulseAgentAssetMappingStatus;
 import com.hivemq.pulse.converters.PulseAgentAssetMappingStatusConverter;
 import com.hivemq.pulse.converters.PulseAgentAssetSchemaConverter;
-import com.hivemq.pulse.status.PulseAgentStatus;
+import com.hivemq.edge.pulse.integration.api.management.PulseAgentStatus;
 import jakarta.ws.rs.core.Response;
 import java.util.List;
 import java.util.UUID;
